pyproject.toml
[project]
name = "ctool"
version = "0.1.0"
description = ""
authors = [
    {name = "Chan Kim",email = "5281995+gogochan@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "elasticsearch (>=8.17.1,<9.0.0)",
    "progressbar2 (>=4.5.0,<5.0.0)",
]

[project.scripts]
ctool = "cli.main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
